requirements.txt
Appium-Python-Client==0.40
assertpy==0.14
ConfigParser==3.7.1
Faker==1.0.2
gherkin-official==4.1.3
py==1.7.0
pylint==2.2.2
pytest==4.2.0
pytest-base-url==1.4.1
pytest-bdd==3.1.0
-e git://github.com/popescunsergiu/pytest-selenium.git@master#egg=pytest_selenium
pytest-xdist==1.26.1
python-dateutil==2.7.5
pyyaml==3.13
requests==2.21.0
selenium==3.141.0
